requirements_3d.txt
accelerate
bitsandbytes==0.35.4
dearpygui
decord==0.6.0
diffusers[torch]==0.31.0
einops>=0.7.0
ftfy
gradio
h5py>=3.10.0
huggingface-hub>=0.22.0
icecream==2.1.0
imageio
imageio-ffmpeg
kiui >= 0.2.3
lpips
matplotlib
modelcards
nerfacc==0.5.3
numpy==1.26.4
omegaconf==2.2.3
onnxruntime
opencv-python
packaging
pandas
peft
pillow>=10.2.0
piq
plyfile
protobuf>=4.25.3
pygltflib
pyhocon==0.3.57
PyMCubes==0.1.4
pyransac3d
pytorch-lightning<2
rembg
rich
roma
safetensors
scikit-image
scikit-learn
scipy
segment_anything
tables
tabulate
tensorboard
termcolor
timm
torch_efficient_distloss
tqdm
transformers==4.51.3
trimesh
tyro
wandb
xatlas
pymeshlab
warp-lang==1.6.2